[BLINK-18838837] [table] Fix ClassCastException when temporal join on an int constant
Summary: fix T18838837

Reviewers: 海豹, 犹夏, 玉兆

Reviewed By: 玉兆

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D841057

+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/table/api/stream/SinkTest.scala
@@ -238,7 +238,7 @@ class SinkTest(subplanReuseEnabled: Boolean)  extends TableTestBase {
 
   @Test
   def testMultiSinksWithTemporalTableSource(): Unit = {
-    util.addTable[(Int, Double, Int)](
+    util.addTable[(Long, Double, Long)](
       "MyTable", 'a, 'b, 'c, 'proctime.proctime, 'rowtime.rowtime)
     util.tableEnv.registerTableSource("TemporalSource", new TestingTemporalTableSource)
 
@@ -435,7 +435,7 @@ class SinkTest(subplanReuseEnabled: Boolean)  extends TableTestBase {
 
   @Test
   def testSingleSinkWithTemporalTableSource(): Unit = {
-    util.addTable[(Int, Double, Int)](
+    util.addTable[(Long, Double, Long)](
       "MyTable", 'a, 'b, 'c, 'proctime.proctime, 'rowtime.rowtime)
     util.tableEnv.registerTableSource("TemporalSource", new TestingTemporalTableSource)
